This is a simple but working version of a products page of the Euro Ceram website
The first two section just displays a bunch of doors and windows that
the user can buy, But the third section(tools) is more dynamic; By
clicking on one of the tags the images change to match the tag title, eg: displays hammers images if the tag is hammer(s).

despite all the work done it still remains the furniture section to be
added.